refactor(react-icons-file-type): consolidate icon src resolvers and dedupe props
- Merge the misnamed getFileTypeIconProps.ts (which contained no
  getFileTypeIconProps) into its only consumer, getFileTypeIconSrc.ts, so all
  URL-resolution logic lives in one cohesive module; delete the old file.
- Fix a duplicated FileTypeIconProps interface declaration in FileTypeIcon.tsx
  that broke compilation.
- Dedupe the props: FileTypeIconProps now extends FileTypeIconOptions instead of
  redeclaring extension/type/size/imageFileType.
- Consolidate the resolver tests into getFileTypeIconSrc.test.ts and remove the
  old getFileTypeIconProps.test.ts.

All symbols are internal (not re-exported from index.ts); public API unchanged.
37 tests pass, lint and build clean.
